Program: * added and managed input parameters -o[output], -f[format: xml/html], -l[list options].
SprinklerTest:
* test are now filtered by category or code, _then_ instantiated.
* log changed to collect results in an instance of TestResults
* some refactory

TestResult:
property Exception is TestFailedException which is now xml-serializable.

(+)TestResults:
new class which holds each testResult instance and can be serialized to xml. New property: ShowStatusInConsole; when true, each time a result is added, a status will be shown in Console.out.

(+)xmlToHtml.xslt:
xsl-transforms an xml-serialized output of TestResults to an Html document.

TestSets:
* gets the list of tests without
* added method GetTestModules
* refactory